backend/campoenorden_backend/requirements.txt
Django
djangorestframework
mysqlclient
Pillow
django-cors-headers